LoongArch: KVM: Repair the restart pressure test virtual machine hang dead problem
Upstream: no

After the virtual restart, exintc does not correctly clear
the estat register and qemu does not clear the estat register.
Estat data sent from qemu is merged instead of overwritten on kvm.

In order to solve this problem,
the following changes were made to this patch:

1.exintc correctly clears zeros when the VM restarts
2. estat on the qemu side clears ESTAT correctly when the VM restarts
3. When the estat register data on the qemu side is sent to the kvm,
the data needs to be overwritten instead of merged

+++ arch/loongarch/kvm/intc/extioi.c
@@ -936,7 +936,7 @@ int kvm_loongarch_reset_extioi(struct kvm *kvm)
 {
 	struct loongarch_extioi *extioi = kvm->arch.extioi;
 	unsigned long flags;
-	u8 offset, size;
+	unsigned long offset, size;
 	u8 *pstart;
 
 	if (!extioi)
